+++ airflow-core/src/airflow/example_dags/example_task_group.py
@@ -24,7 +24,7 @@ import pendulum
 from airflow.providers.standard.operators.bash import BashOperator
 from airflow.providers.standard.operators.empty import EmptyOperator
 from airflow.sdk import DAG
-from airflow.utils.task_group import TaskGroup
+from airflow.sdk.definitions.taskgroup import TaskGroup
 
 # [START howto_task_group]
 with DAG(
