+++ airflow-core/src/airflow/utils/__init__.py
@@ -46,5 +46,11 @@ __deprecated_classes = {
     "xcom": {
         "XCOM_RETURN_KEY": "airflow.models.xcom.XCOM_RETURN_KEY",
     },
+    "task_group": {
+        "TaskGroup": "airflow.sdk.definitions.taskgroup.TaskGroup",
+        "MappedTaskGroup": "airflow.sdk.definitions.taskgroup.MappedOperator",
+        "get_task_group_children_getter": "airflow.sdk.definitions.taskgroup.get_task_group_children_getter",
+        "task_group_to_dict": "airflow.sdk.definitions.taskgroup.task_group_to_dict",
+    },
 }
 add_deprecated_classes(__deprecated_classes, __name__)
