+++ airflow-core/tests/unit/utils/test_dag_cycle.py
@@ -22,8 +22,8 @@ from airflow.exceptions import AirflowDagCycleException
 from airflow.models.dag import DAG
 from airflow.providers.standard.operators.empty import EmptyOperator
 from airflow.sdk import Label
+from airflow.sdk.definitions.taskgroup import TaskGroup
 from airflow.utils.dag_cycle_tester import check_cycle
-from airflow.utils.task_group import TaskGroup
 
 from unit.models import DEFAULT_DATE
